Update Bank problem.cpp

+++ Oops questions practice -01/Bank problem.cpp
@@ -5,7 +5,7 @@ class BankAccount
     private :
     
     string bankname;
-    int accountNumber;
+    double accountNumber;
     double balance;
     public :
